pushed UI improvements
