+++ src/app/(game)/pmwgame/PMWGame.tsx
@@ -1,34 +1,25 @@
 "use client";
-import {
-  Text,
-  PixelBorder,
-  Progress,
-  Button,
-  Input,
-  Toolbar,
-  Separator,
-  Container,
-  Table,
-} from "nes-ui-react";
+import { Text, Progress, Button, Input, Container, Table } from "nes-ui-react";
 import { useAccount } from "wagmi";
 import Grid from "../Components/Grid";
 import { useEffect, useRef, useState } from "react";
-import { useConnectModal } from "@rainbow-me/rainbowkit";
 import useEnterGame from "@/hooks/sc-fns/useEnterGame";
 import useEliminateUser from "@/hooks/sc-fns/useEliminateUser";
-import { UserWalletAddress } from "./UserWalletAddress";
 import usePMWReader from "@/hooks/sc-fns/usePMWReader";
 import useUserTickets from "@/hooks/sc-fns/useUserTickets";
 import useCurrentRound from "@/hooks/sc-fns/useCurrentRound";
 import usePrizes from "@/hooks/sc-fns/usePrizes";
 import { extractRoundsPrizes } from "../utils";
 import JackpotDisplay from "./JackpotDisplay";
+import Header from "../Components/Header";
+import PMWTitle from "../Components/PMWTitle";
+import ConnectWallet from "../Components/ConnectWallet";
+import EnterGameScreen from "../Components/EnterGameScreen";
 
 type PrizeArrays = [bigint[], bigint[], boolean[], string[], bigint[]];
 
 export default function PMWGame() {
   const account = useAccount();
-  const { openConnectModal } = useConnectModal();
   const { PMWReader, PMWError, PMWFetched, refetchPMWReader } = usePMWReader();
 
   const [isRegistrationOpen, setIsRegistrationOpen] = useState(true);
@@ -180,17 +171,6 @@ export default function PMWGame() {
   const userAllTicketsCount = (userAllTickets as [number[], boolean[]])?.[0]
     ?.length;
 
-  const handleLogin = (e: React.MouseEvent) => {
-    e.preventDefault();
-    if (openConnectModal) {
-      openConnectModal();
-    }
-  };
-
-  function handleChange(value: string) {
-    setUserTickets(value === "" ? 0 : Number(value));
-  }
-
   const {
     enterGame,
     transactionLoading: isEnterGameTransactionLoading,
@@ -238,109 +218,38 @@ export default function PMWGame() {
     }
     return "Mist 'em";
   };
-  return (
-    <div className="m-8 flex flex-col justify-center items-center gap-[1rem] box-border">
-      <PixelBorder
-        doubleSize
-        doubleRoundCorners
-        style={{
-          width: "max-content",
-          padding: "8px 16px",
-          marginBottom: "16px",
-        }}
-      >
-        <Text size="large" centered>
-          <br />
-          Pink Mist Whale
-        </Text>
-      </PixelBorder>
-      <Toolbar
-        style={{
-          padding: "10px 15px",
-          width: "80vw",
-          display: "flex",
-          justifyContent: "space-between",
-          alignItems: "center",
-        }}
-      >
-        <div className="w-[33%] flex justify-center items-center">
-          <Text size="large" style={{ marginBottom: "0" }}>
-            Your Active Tickets: {userActiveTicketCount || 0}
-          </Text>
-        </div>
-        <Separator />
-        <div className="w-[33%] flex justify-center items-center">
-          <Text size="large" style={{ marginBottom: "0" }}>
-            ROUND-{Number(currentRoundId)}
-          </Text>
-        </div>
-        {account.isConnected && (
-          <>
-            <Separator />
-            <div className="w-[33%] flex justify-center items-center">
-              <UserWalletAddress />
-            </div>
-          </>
-        )}
-      </Toolbar>
-      <JackpotDisplay
-        daiAmount={Number((currentRoundPrize as any).daiAmount || 0)}
-        darkAmount={Number((currentRoundPrize as any).darkAmount || 0)}
-      />
-      {account.isConnected ? (
-        <>
-          <div className="flex flex-col items-center">
-            <Text>
-              {isRegistrationOpen
-                ? `${currentActiveTicketsCount} out of 
-            ${totalParticipants}...`
-                : ""}
-            </Text>
-            {isRegistrationOpen && (
-              <Progress
-                value={currentActiveTicketsCount}
-                max={totalParticipants}
-                color="pattern"
-                style={{ width: "40vw" }}
-              />
-            )}
-          </div>
 
+  const isAccountConnected = account.isConnected;
+  return (
+    <div className="min-h-screen p-8 flex flex-col box-border items-center">
+      <PMWTitle />
+      <Header
+        activeTicketsCount={userActiveTicketCount || 0}
+        currentRoundId={Number(currentRoundId)}
+        isAccountConnected={isAccountConnected}
+      />
+      {isAccountConnected && isRegistrationOpen ? (
+        <EnterGameScreen
+          currentRoundPrize={currentRoundPrize}
+          activeTicketsCount={currentActiveTicketsCount}
+          totalParticipants={totalParticipants}
+          userTickets={userTickets}
+          setUserTickets={setUserTickets}
+          renderEnterGameButtonState={renderEnterGameButtonState}
+          maxTickets={maxTickets}
+          isEnterGameTransactionLoading={isEnterGameTransactionLoading}
+          enterGame={enterGame}
+          userAllTicketsCount={userAllTicketsCount}
+          userAllTickets={userAllTickets}
+          lastRoundsPrizes={lastRoundsPrizes}
+          currentRoundId={currentRoundId}
+        />
+      ) : null}
+
+      {isAccountConnected ? (
+        <>
           <div className="flex items-center justify-center">
-            {isRegistrationOpen ? (
-              <>
-                <div className="flex flex-col">
-                  <Input
-                    type="number"
-                    name="userTickets"
-                    value={userTickets.toString()}
-                    label="Number Of Tickets: "
-                    style={{ height: "32px", fontSize: "16px" }}
-                    onChange={handleChange}
-                    color={
-                      userTickets < 0 || userTickets > maxTickets
-                        ? "error"
-                        : "none"
-                    }
-                  />
-                  <Text size="medium" color="warning">
-                    Max Tickets: {maxTickets}
-                  </Text>
-                </div>
-                <Button
-                  color="primary"
-                  size="large"
-                  disabled={
-                    userTickets <= 0 ||
-                    userTickets > maxTickets ||
-                    isEnterGameTransactionLoading
-                  }
-                  onClick={enterGame}
-                >
-                  {renderEnterGameButtonState()}
-                </Button>
-              </>
-            ) : (
+            {!isRegistrationOpen ? (
               <Button
                 color="primary"
                 onClick={eliminateUser}
@@ -349,7 +258,7 @@ export default function PMWGame() {
               >
                 <Text size="large">{renderEliminateUserButtonState()}</Text>
               </Button>
-            )}
+            ) : null}
           </div>
 
           <div className="flex w-full max-w-full box-border gap-[8px]">
@@ -360,88 +269,11 @@ export default function PMWGame() {
                   activeTicketCount={currentActiveTicketsCount}
                 />
               </div>
-            ) : (
-              <div>
-                {lastRoundsPrizes.length !== 0
-                  ? `Congratulations ${lastRoundsPrizes?.[0]?._winner} for winning
-              Round-${Number(currentRoundId) - 1}`
-                  : null}
-              </div>
-            )}
-            <div className="w-[25%] max-w-[25%] flex flex-col gap-[20px]">
-              <Container
-                align="left"
-                title="&lt;Your Tickets&gt;"
-                roundedCorners
-                alignTitle="center"
-                style={{ width: "fit-content" }}
-              >
-                <div className="flex gap-[10px] mt-5 items-center justify-start">
-                  {userAllTicketsCount === 0 ? (
-                    <span className="!text-[16px] !text-pretty">
-                      {isRegistrationOpen
-                        ? "Hit 'Enter Game' to buy Tickets! "
-                        : "Mist 'em all to participate in the next round!"}
-                    </span>
-                  ) : (
-                    (userAllTickets as [number[], boolean[]])?.[0]?.map(
-                      (ticket, index) => {
-                        return (
-                          <span
-                            key={index}
-                            className={`${
-                              (userAllTickets as [number[], boolean[]])?.[1]?.[
-                                index
-                              ]
-                                ? "text-successgreen"
-                                : "text-brred"
-                            } !text-[16px]`}
-                          >
-                            {Number(ticket) + 1}
-                          </span>
-                        );
-                      },
-                    )
-                  )}
-                </div>
-              </Container>
-              {lastRoundsPrizes.length !== 0 ? (
-                <Table>
-                  <thead>
-                    <tr>
-                      <th>ROUND ID</th>
-                      <th>WINNER</th>
-                      <th>AMOUNT WON</th>
-                      <th>TICKET NUMBER</th>
-                    </tr>
-                  </thead>
-                  <tbody>
-                    {lastRoundsPrizes?.map((roundPrize) => {
-                      return (
-                        <tr key={roundPrize?.roundId}>
-                          <td>{roundPrize?.roundId}</td>
-                          <td>
-                            0x...
-                            {String(roundPrize?._winner)?.slice(-3)}
-                          </td>
-                          <td>{`${Number(roundPrize?.daiAmount)} $DAI + ${Number(roundPrize?.darkAmount)} $DARK`}</td>
-                          <td>{Number(roundPrize?.winningTicket)}</td>
-                        </tr>
-                      );
-                    })}
-                  </tbody>
-                </Table>
-              ) : null}
-            </div>
+            ) : null}
           </div>
         </>
       ) : (
-        <div className="flex flex-col items-center justify-center h-[100%] mt-[100px]">
-          <Text size="large">Please Connect Your Wallet to Play The Game.</Text>
-          <Button color="primary" size="large" onClick={handleLogin}>
-            Connect Wallet
-          </Button>
-        </div>
+        <ConnectWallet />
       )}
     </div>
   );
